+++ src/main/index.ts
@@ -10,12 +10,10 @@ let mainWindow: BrowserWindow | null;
 function createMainWindow() {
   const window = new BrowserWindow({ webPreferences: { nodeIntegration: true } });
   window.maximize();
+  window.removeMenu();
 
   if (isDevelopment) {
     window.webContents.openDevTools();
-  }
-
-  if (isDevelopment) {
     window.loadURL(`http://localhost:${process.env.ELECTRON_WEBPACK_WDS_PORT}`);
   } else {
     window.loadURL(