+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This changelog keeps track of most (except early development) changes to the too
 ### Updated
 
 - No longer need to reference `xoreos-tools` directory, tools are bundled in
+- Remove default electron menu, keep dev tools shortcut + add a nice debug icon (just in case we need logs)
 
 ### Fixed